+++ apps/xim2_web/lib/xim2_web/controllers/page_html/colors.html.heex
@@ -1,5 +1,7 @@
 <.main_title>Colors</.main_title>
 
+<.back navigate={~p"/"}>Home</.back>
+
 <table class="mb-3 p-3 bg-neutral-500">
   <tr>
     <td class="bg-slate-50 w-40 h-10 text-center">slate</td>
