+++ apps/xim2_web/lib/xim2_web/controllers/page_html/home.html.heex
@@ -1,2 +1,15 @@
+<.header_visual
+  small={~p"/images/blue-world-map-sm.jpg"}
+  medium={~p"/images/blue-world-map-md.jpg"}
+  large={~p"/images/blue-world-map-lg.jpg"}
+/>
+
 <.flash_group flash={@flash} />
 <.main_title>Home</.main_title>
+
+<p>
+  <ul>
+    <li><.link class="text-sky-200" navigate={~p"/grid"}>Grid</.link></li>
+    <li><.link class="text-sky-200" navigate={~p"/colors"}>Colors</.link></li>
+  </ul>
+</p>
